refactor(test) : 테스트 버그 수정
