+++ src/test/java/kr/hhplus/be/server/application/order/InventoryServiceTest.java
@@ -1,6 +1,6 @@
 package kr.hhplus.be.server.application.order;
 
-import kr.hhplus.be.server.application.order.service.InventoryService;
+import kr.hhplus.be.server.application.product.InventoryService;
 import kr.hhplus.be.server.domain.order.entity.Order;
 import kr.hhplus.be.server.domain.order.entity.OrderItem;
 import kr.hhplus.be.server.domain.order.service.OrderService;
