+++ src/test/java/kr/hhplus/be/server/application/ranking/DailyRankingIntegrationTest.java
@@ -1,7 +1,6 @@
 package kr.hhplus.be.server.application.ranking;
 
-import kr.hhplus.be.server.application.order.service.CompensationService;
-import kr.hhplus.be.server.application.order.service.InventoryService;
+import kr.hhplus.be.server.application.product.InventoryService;
 import kr.hhplus.be.server.domain.balance.entity.Balance;
 import kr.hhplus.be.server.domain.balance.repository.BalanceRepository;
 import kr.hhplus.be.server.domain.order.entity.Order;
@@ -66,9 +65,6 @@ class DailyRankingIntegrationTest {
 
     @MockitoBean
     private InventoryService inventoryService;
-
-    @MockitoSpyBean
-    private CompensationService compensationService;
 
     private User testUser;
     private Product p1, p2;
